requirements.txt
google.generativeai == 0.6.0
chromadb == 0.5.0
pypdf == 4.2.0